+++ corpus/labeled/hn_guarded_json.js
@@ -1,0 +1,8 @@
+async function load(res) {
+  try {
+    return JSON.parse(await res.text());
+  } catch (e) {
+    return null;
+  }
+}
+module.exports = load;