+++ corpus/labeled/tp_debugger.js
@@ -1,0 +1,5 @@
+function f() {
+  debugger;
+  return 1;
+}
+module.exports = f;