+++ corpus/labeled/tp_emptycatch.js
@@ -1,0 +1,6 @@
+function f() {
+  try {
+    risky();
+  } catch (e) {}
+}
+module.exports = f;