+++ corpus/labeled/tp_eval.js
@@ -1,0 +1,4 @@
+function run(userInput) {
+  return eval(userInput);
+}
+module.exports = run;